TrackSelectionHelper.java:
+ simplified logic for selecting the preferred audio/text tracks.
+ setters renamed to give more accurate name
+ PrefferedSelectionMode renamed to Mode and moved to PKTrackConfig.java
- deleted PKTrackLanguage.java

+++ playkit/src/main/java/com/kaltura/playkit/player/MediaPlayerWrapper.java
@@ -267,7 +267,7 @@ class MediaPlayerWrapper implements PlayerEngine, SurfaceHolder.Callback, MediaP
 
     @Override
     public long getBufferedPosition() {
-        return Double.valueOf(Math.floor(playerDuration * (currentBufferPercentage / Consts.PERCENT_FACTOR))).longValue();
+        return (long) Math.floor(playerDuration * (currentBufferPercentage / Consts.PERCENT_FACTOR));
     }
 
     @Override
